requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
mongoengine==0.28.2
pymongo==4.7.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
